// packages/ds/angular/src/lib/jl-cl/jl-cl/validator-types.ts
export enum ErrorNames {
    required = 'required',
    invalid = 'invalid',
    pattern = 'pattern',
    disallowed = 'disallowed character',
    minDate = 'minimum date',
    maxDate = 'maximum date',
    minLength = 'minlength',
    maxLength = 'maxlength',
    unmatched = 'unmatched',
    tableRequired = 'tableRequired',
    invalidDate = 'invalidDate',
    primaryNumberRequired = 'primarNumberRequired',
}